video routes, profile routes protected

+++ controllers/videoController.js
@@ -38,8 +38,6 @@ export const postVideoUpload = async (req, res) => {
     description,
     creator: req.user.id,
   });
-  console.log(newVideo);
-  console.log(newVideo.id);
   req.user.videos.push(newVideo.id);
   req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
@@ -51,7 +49,6 @@ export const videoDetail = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id).populate("creator");
-    console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -64,7 +61,11 @@ export const getVideoEdit = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("videoEdit", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("videoEdit", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -88,9 +89,13 @@ export const videoDelete = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
-    console.log(error);
+    res.redirect(routes.home);
   }
-  res.redirect(routes.home);
 };
